s l a s h   c o m m a n d s

+++ index.js
@@ -16,6 +16,7 @@ let waiting = []
 let phrases = ["gib cookie"]
 let config = {}
 let commands = []
+let gcommands = []
 let emojis = []
 
 async function loadPhrases() {
@@ -44,12 +45,18 @@ async function loadPhrases() {
     }catch(e) {
         console.error("Failed to parse blacklisted people! D:")
     }
+    try {
+        gcommands = JSON.parse(await fs.readFile(__dirname + "/gcommands.json"))
+    }catch(e) {
+        console.error("Failed to parse global commands! D:")
+    }
 }
 
 bot.on("message",async (m) => {
     if(m.author.bot) return;
     if(m.content.toLowerCase().startsWith("🍪unblacklist") || m.content.toLowerCase().startsWith("🍪 unblacklist")) {
         if(!config.admins.includes(m.author.id)) return;
+        if(!m.content.split(">")[1]) return;
         waiting.push(m.content.split(">")[1])
         var current = await fs.readFile(__dirname + "/blacklisted.json")
         var arr = JSON.parse(current)
@@ -94,12 +101,30 @@ bot.on("message",async (m) => {
         for(var c of commands) {
             await bot.api.applications(bot.user.id).guilds(m.guild.id).commands.post({data: c})
         }
+        for(var c2 of gcommands) {
+            await bot.api.applications(bot.user.id).commands.post({data: c2})
+        }
         console.log(await bot.api.applications(bot.user.id).guilds(m.guild.id).commands.get())
+        console.log(await bot.api.applications(bot.user.id).commands.get())
         m.channel.send("[🍪] Added commands! :D")
         return
     }
+    if(m.content.toLowerCase() == "🍪debug" || m.content.toLowerCase() == "🍪 debug") {
+        var list = await bot.api.applications(bot.user.id).guilds(m.guild.id).commands.get()
+        var glist = await bot.api.applications(bot.user.id).commands.get()
+        var ms = "**GUILD**\n"
+        list.forEach(c => {
+            ms += `${c.name} | \`${c.id}\`\n`
+        })
+        ms += "\n**GLOBAL**\n"
+        glist.forEach(c => {
+            ms += `${c.name} | \`${c.id}\`\n`
+        })
+        return m.channel.send(ms)
+    }
     if(m.content.toLowerCase().startsWith("🍪blacklist") || m.content.toLowerCase().startsWith("🍪 blacklist")) {
         if(!config.admins.includes(m.author.id)) return;
+        if(!m.content.split(">")[1]) return;
         waiting.push(m.content.split(">")[1])
         var current = await fs.readFile(__dirname + "/blacklisted.json")
         var arr = JSON.parse(current)
@@ -110,7 +135,16 @@ bot.on("message",async (m) => {
     }
     if(m.content.toLowerCase().startsWith("🍪delc") || m.content.toLowerCase().startsWith("🍪 delc")) {
         if(!config.admins.includes(m.author.id)) return;
-        await bot.api.applications(bot.user.id).guilds(m.guild.id).commands(m.content.split(">")[1]).delete()
+        if(!m.content.split(">")[1]) return;
+        try {
+            await bot.api.applications(bot.user.id).guilds(m.guild.id).commands(m.content.split(">")[1]).delete()
+        }catch(e) {
+            try {
+                await bot.api.applications(bot.user.id).commands(m.content.split(">")[1]).delete()
+            }catch(e) {
+                m.channel.send("[🍪] failed :(")
+            }
+        }
         m.channel.send("[🍪] Successfully deleted command! :/")
         return
     }
@@ -127,33 +161,82 @@ bot.on("message",async (m) => {
 })
 
 bot.ws.on('INTERACTION_CREATE', async interaction => {
-    if(waiting.includes(interaction.member.user.id)) return;
+    if(waiting.includes(interaction.member.user.id)) {
+        bot.api.interactions(interaction.id, interaction.token).callback.post({data: 
+            {
+                type: 3,
+                data: {
+                    content: `Cookie Bot is upset,\nyou are requesting too many cookies too quickly >:(`,
+                    flags: 1 << 6
+                }
+            }
+        })
+        return;
+    }
     waiting.push(interaction.member.user.id)
     setTimeout(() => {
         waiting = waiting.filter(e => e !== interaction.member.user.id)
     },5000)
-    if(!interaction?.data?.options || !interaction?.data?.options?.length) {
+    if(interaction.data.name.toLowerCase() == "cookie") {
+        if(!interaction?.data?.options || !interaction?.data?.options?.length) {
+            await bot.api.interactions(interaction.id, interaction.token).callback.post({data: 
+                {
+                    type: 3,
+                    data: {
+                        content: `Here's your cookie, <@${interaction.member.user.id}>!`
+                    }
+                }
+            })
+            bot.channels.cache.get(interaction.channel_id).send("🍪")
+            return;
+        }
+        // do stuff and respond here
         await bot.api.interactions(interaction.id, interaction.token).callback.post({data: 
             {
                 type: 3,
                 data: {
-                    content: `Here's your cookie, <@${interaction.member.user.id}>!`
+                    content: interaction.data?.options[1]?.value ? `<@${interaction.data.options[0].value}>, someone gave you a cookie!` : `<@${interaction.data.options[0].value}>, you got a cookie from <@${interaction.member.user.id}>!`,
+                    // flags: 1 << 6 
                 }
             }
         })
         bot.channels.cache.get(interaction.channel_id).send("🍪")
-        return;
-    }
-    // do stuff and respond here
-    await bot.api.interactions(interaction.id, interaction.token).callback.post({data: 
-        {
-            type: 3,
-            data: {
-                content: interaction.data?.options[1]?.value ? `<@${interaction.data.options[0].value}>, someone gave you a cookie!` : `<@${interaction.data.options[0].value}>, you got a cookie from <@${interaction.member.user.id}>!` 
-            }
+    }else if(interaction.data.name.toLowerCase() == "statement") {
+        // this is protected >:(
+        if(!config.admins.includes(interaction.member.user.id)) {
+            bot.api.interactions(interaction.id, interaction.token).callback.post({data: 
+                {
+                    type: 3,
+                    data: {
+                        content: `Cookie Bot is upset,\nyou aren't supposed to use this command!!!`,
+                        flags: 1 << 6
+                    }
+                }
+            })
+            return;
+        };
+        if(interaction.data.options) {
+            await bot.api.interactions(interaction.id, interaction.token).callback.post({data: 
+                {
+                    type: 3,
+                    data: {
+                        content: `> ${interaction.data.options[0].value}\n - <@${bot.user.id}>`,
+                        // flags: 1 << 6 
+                    }
+                }
+            })
+        }else{
+            await bot.api.interactions(interaction.id, interaction.token).callback.post({data: 
+                {
+                    type: 3,
+                    data: {
+                        content: `> Cookies are cool\n - <@${bot.user.id}>`,
+                        // flags: 1 << 6 
+                    }
+                }
+            })
         }
-    })
-    bot.channels.cache.get(interaction.channel_id).send("🍪")
+    }
 })
 
 process.on("SIGINT",async () => {
